+++ src/cmd/5c_/txt.c
@@ -147,7 +147,7 @@ nextpc(void)
 void
 gargs(Node *n, Node *tn1, Node *tn2)
 {
-	long regs;
+	int32 regs;
 	Node fnxargs[20], *fnxp;
 
 	regs = cursafe;
@@ -217,7 +217,7 @@ garg1(Node *n, Node *tn1, Node *tn2, int f, Node **fnxp)
 }
 
 Node*
-nodconst(long v)
+nodconst(int32 v)
 {
 	constnode.vconst = v;
 	return &constnode;
@@ -448,7 +448,7 @@ raddr(Node *n, Prog *p)
 void
 naddr(Node *n, Adr *a)
 {
-	long v;
+	int32 v;
 
 	a->type = D_NONE;
 	if(n == Z)
@@ -1176,7 +1176,7 @@ gbranch(int o)
 }
 
 void
-patch(Prog *op, long pc)
+patch(Prog *op, int32 pc)
 {
 
 	op->to.offset = pc;
@@ -1224,7 +1224,7 @@ sconst(Node *n)
 }
 
 int
-sval(long v)
+sval(int32 v)
 {
 	int i;
 
@@ -1233,7 +1233,7 @@ sval(long v)
 			return 1;
 		if((~v & ~0xff) == 0)
 			return 1;
-		v = (v<<2) | ((ulong)v>>30);
+		v = (v<<2) | ((uint32)v>>30);
 	}
 	return 0;
 }
@@ -1241,7 +1241,7 @@ sval(long v)
 int32
 exreg(Type *t)
 {
-	long o;
+	int32 o;
 
 	if(typechlp[t->etype]) {
 		if(exregoffset <= REGEXT-2)
